centered data elements of form
add header to form table with dates
