--- zootable/zoo_checks/templatetags/index.py
@@ -1,8 +1,0 @@
-from django import template
-
-register = template.Library()
-
-
-@register.filter
-def index(List, i):
-    return List[int(i)]
